fix(web): add prose styling for README markdown + quality cleanup
- Install @tailwindcss/typography so prose classes actually style
  the rendered markdown (headings, lists, code blocks, links)
- Extract triplicated relativeTime() into shared/lib/relative-time.ts
- Remove debug console.log/error statements from git clone hook
