+++ package.json
@@ -7,5 +7,8 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.4.6"
+  },
+  "dependencies": {
+    "@tailwindcss/typography": "^0.5.19"
   }
 }
