+++ web/src/features/repos/ui/RepoListItem.tsx
@@ -3,30 +3,12 @@ import { Link } from "@tanstack/react-router";
 
 import { Badge } from "@/shared/ui/badge";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/shared/ui/tooltip";
+import { relativeTime } from "@/shared/lib/relative-time";
 import type { Repo } from "../use-repos";
 
 function truncateHex(hex: string): string {
   if (hex.length <= 12) return hex;
   return `${hex.slice(0, 8)}...${hex.slice(-4)}`;
-}
-
-function relativeTime(unix: number): string {
-  const now = Date.now();
-  const diff = now - unix * 1000;
-  const seconds = Math.floor(diff / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-
-  if (days > 30) {
-    const months = Math.floor(days / 30);
-    return months === 1 ? "1 month ago" : `${months} months ago`;
-  }
-  if (days > 0) return days === 1 ? "1 day ago" : `${days} days ago`;
-  if (hours > 0) return hours === 1 ? "1 hour ago" : `${hours} hours ago`;
-  if (minutes > 0)
-    return minutes === 1 ? "1 minute ago" : `${minutes} minutes ago`;
-  return "just now";
 }
 
 export function RepoListItem({ repo }: { repo: Repo }) {
